tests/wpt/web-platform-tests/web-animations/responsive/borderWidth.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id='container'>
<div id='element'></div>
</div>

<script>
var element = document.getElementById('element');
var container = document.getElementById('container');

test(function() {
    element.style.fontSize = '10px';
    element.style.border = 'none';
    var player = element.animate([{borderBottomWidth: '300px'}, {borderBottomWidth: '10em'}], 10);
    player.pause();
    player.currentTime = 5;
    element.style.border = 'solid';
    assert_equals(getComputedStyle(element).borderBottomWidth, '200px');
}, 'Border width responsive to border style changes');

test(function() {
    element.style.fontSize = '50px';
    element.style.border = 'solid';
    var player = element.animate([{borderBottomWidth: '300px'}, {borderBottomWidth: '10em'}], 10);
    player.pause();
    player.currentTime = 5;
    element.style.fontSize = '10px';
    assert_equals(getComputedStyle(element).borderBottomWidth, '200px');
}, 'Border width responsive to font size changes');

</script>
